// src/components/Catalog/CatalogItem/CatalogItem.types.ts
export interface ICar {
  accessories: string[];
  address: string;
  description: string;
  engineSize: string;
  fuelConsumption: string;
  functionalities: string[];
  id: number;
  img: string;
  make: string;
  mileage: number;
  model: string;
  rentalCompany: string;
  rentalConditions: string;
  rentalPrice: string;
  type: string;
  year: number;
}